+++ test/util/framework/hcp_helper.go
@@ -29,18 +29,13 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/davecgh/go-spew/spew"
-	"github.com/go-logr/logr"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/onsi/ginkgo/v2"
 	"golang.org/x/crypto/ssh"
-	"golang.org/x/sync/errgroup"
 
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -48,13 +43,10 @@ import (
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 
 	"github.com/Azure/ARO-HCP/internal/api"
 	hcpsdk20240610preview "github.com/Azure/ARO-HCP/test/sdk/resourcemanager/redhatopenshifthcp/armredhatopenshifthcp"
 	hcpsdk20251223preview "github.com/Azure/ARO-HCP/test/sdk/v20251223preview/resourcemanager/redhatopenshifthcp/armredhatopenshifthcp"
-	hcpsdk20260630preview "github.com/Azure/ARO-HCP/test/sdk/v20260630preview/resourcemanager/redhatopenshifthcp/armredhatopenshifthcp"
 )
 
 // checkOperationResult ensures the result model returned by a runtime.Poller
@@ -79,286 +71,6 @@ func checkOperationResult(expectModel, resultModel any) error {
 	}
 
 	return nil
-}
-
-func (tc *perItOrDescribeTestContext) GetAdminRESTConfigForHCPCluster20240610(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	timeout time.Duration, // this is a POST request, so keep the timeout as it's async
-) (*rest.Config, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during GetAdminRESTConfigForHCPCluster for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	startTime := time.Now()
-	defer func() {
-		finishTime := time.Now()
-		tc.RecordTestStep("Collect admin credentials for cluster", startTime, finishTime)
-	}()
-
-	adminCredentialRequestPoller, err := hcpClient.BeginRequestAdminCredential(
-		ctx,
-		resourceGroupName,
-		hcpClusterName,
-		nil,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to start credential request: %w", err)
-	}
-
-	operationResult, err := adminCredentialRequestPoller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return nil, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish getting creds, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-		}
-		return nil, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish getting creds: %w", hcpClusterName, resourceGroupName, err)
-	}
-
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.HcpOpenShiftClustersClientRequestAdminCredentialResponse:
-		restConfig, err := clientcmd.BuildConfigFromKubeconfigGetter("", func() (*clientcmdapi.Config, error) {
-			if m.Kubeconfig == nil {
-				return nil, fmt.Errorf("kubeconfig content is nil")
-			}
-			return clientcmd.Load([]byte(*m.Kubeconfig))
-		})
-		if err != nil {
-			return nil, err
-		}
-
-		tc.contextLock.Lock()
-		tc.hcpAdminConfigs[resourceGroupName+"/"+hcpClusterName] = restConfig
-		tc.contextLock.Unlock()
-
-		return restConfig, nil
-	default:
-		return nil, fmt.Errorf("unknown type %T", m)
-	}
-}
-
-func (tc *perItOrDescribeTestContext) RevokeCredentialsAndWait20240610(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	timeout time.Duration,
-) error {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during RevokeCredentialsAndWait for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	startTime := time.Now()
-	defer func() {
-		finishTime := time.Now()
-		tc.RecordTestStep("Collect revoke admin credentials for cluster", startTime, finishTime)
-	}()
-
-	poller, err := hcpClient.BeginRevokeCredentials(ctx, resourceGroupName, hcpClusterName, nil)
-	if err != nil {
-		return fmt.Errorf("failed to start credential revocation for hcpCluster=%q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-	}
-
-	operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish revoking creds, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-		}
-		return fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish revoking creds: %w", hcpClusterName, resourceGroupName, err)
-	}
-
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.HcpOpenShiftClustersClientRevokeCredentialsResponse:
-		return nil
-	default:
-		return fmt.Errorf("unknown type %T", m)
-	}
-}
-
-// DeleteHCPCluster20240610 deletes an hcp cluster and waits for the operation to complete
-// Transient ConflictingConcurrentWriteNotAllowed (409) errors are retried automatically
-// with exponential backoff.
-func DeleteHCPCluster20240610(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	timeout time.Duration,
-) error {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during DeleteHCPCluster for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	var attempt int
-	var lastTransientErr error
-	retryErr := wait.ExponentialBackoffWithContext(ctx, stateConflictBackoff, func(ctx context.Context) (bool, error) {
-		attempt++
-		err := deleteHCPClusterAttempt(ctx, hcpClient, resourceGroupName, hcpClusterName)
-		if err == nil {
-			return true, nil
-		}
-		if isTransientDeleteError(err) {
-			lastTransientErr = err
-			ginkgo.GinkgoLogr.Info("transient conflict during cluster delete, retrying",
-				"cluster", hcpClusterName, "resourceGroup", resourceGroupName,
-				"attempt", attempt, "error", err.Error())
-			return false, nil
-		}
-		return false, err
-	})
-
-	if wait.Interrupted(retryErr) && lastTransientErr != nil {
-		return fmt.Errorf("delete interrupted for hcpCluster=%q in resourcegroup=%q after %d attempts, last transient error: %w, interrupt cause: %w", hcpClusterName, resourceGroupName, attempt, lastTransientErr, retryErr)
-	}
-
-	return retryErr
-}
-
-// deleteHCPClusterAttempt performs a single delete attempt for an HCP cluster.
-func deleteHCPClusterAttempt(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-) error {
-	poller, err := hcpClient.BeginDelete(ctx, resourceGroupName, hcpClusterName, nil)
-	if err != nil {
-		var respErr *azcore.ResponseError
-		if errors.As(err, &respErr) && respErr.StatusCode == http.StatusConflict {
-			resp, getErr := hcpClient.Get(ctx, resourceGroupName, hcpClusterName, nil)
-			if getErr == nil && resp.Properties != nil && resp.Properties.ProvisioningState != nil && *resp.Properties.ProvisioningState == hcpsdk20240610preview.ProvisioningStateDeleting {
-				ginkgo.GinkgoLogr.Info("cluster already deleting, waiting for completion",
-					"cluster", hcpClusterName, "resourceGroup", resourceGroupName)
-				return waitForHCPClusterDeletion20240610(ctx, hcpClient, resourceGroupName, hcpClusterName)
-			}
-		}
-		return err
-	}
-
-	operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish deleting, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-		}
-		return fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish deleting: %w", hcpClusterName, resourceGroupName, err)
-	}
-
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.HcpOpenShiftClustersClientDeleteResponse:
-	default:
-		fmt.Printf("#### unknown type %T: content=%v", m, spew.Sdump(m))
-		return fmt.Errorf("unknown type %T", m)
-	}
-
-	return nil
-}
-
-// waitForHCPClusterDeletion20240610 polls GET on the cluster until it returns 404 (deleted).
-func waitForHCPClusterDeletion20240610(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-) error {
-	for {
-		_, err := hcpClient.Get(ctx, resourceGroupName, hcpClusterName, nil)
-		if err != nil {
-			var respErr *azcore.ResponseError
-			if errors.As(err, &respErr) && respErr.StatusCode == http.StatusNotFound {
-				ginkgo.GinkgoLogr.Info("cluster deletion completed",
-					"cluster", hcpClusterName, "resourceGroup", resourceGroupName)
-				return nil
-			}
-			if errors.Is(err, context.DeadlineExceeded) {
-				return fmt.Errorf("timed out waiting for already-deleting hcpCluster=%q in resourcegroup=%q to be deleted, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return fmt.Errorf("failed polling for deletion of hcpCluster=%q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-		}
-
-		select {
-		case <-ctx.Done():
-			return fmt.Errorf("timed out waiting for already-deleting hcpCluster=%q in resourcegroup=%q, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), ctx.Err())
-		case <-time.After(StandardPollInterval):
-		}
-	}
-}
-
-// UpdateHCPCluster20240610 updates an HCP cluster using the v20240610preview SDK and waits for the operation to complete.
-// Transient 500, 409, and CS state conflict 400 errors are retried automatically with exponential backoff.
-func UpdateHCPCluster20240610(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	update hcpsdk20240610preview.HcpOpenShiftClusterUpdate,
-	timeout time.Duration,
-) (*hcpsdk20240610preview.HcpOpenShiftCluster, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during UpdateHCPCluster for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	var hcpOpenShiftCluster *hcpsdk20240610preview.HcpOpenShiftCluster
-	var lastTransientErr error
-	attempt := 0
-
-	err := wait.ExponentialBackoffWithContext(ctx, stateConflictBackoff, func(ctx context.Context) (bool, error) {
-		attempt++
-		if attempt > 1 {
-			ginkgo.GinkgoLogr.Info("Retrying cluster update",
-				"cluster", hcpClusterName,
-				"attempt", attempt)
-		}
-
-		poller, err := hcpClient.BeginUpdate(ctx, resourceGroupName, hcpClusterName, update, nil)
-		if err != nil {
-			if isTransientUpdateError(err) {
-				lastTransientErr = err
-				return false, nil
-			}
-			return false, err
-		}
-
-		operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-			Frequency: StandardPollInterval,
-		})
-		if err != nil {
-			if isTransientUpdateError(err) {
-				lastTransientErr = err
-				return false, nil
-			}
-			if errors.Is(err, context.DeadlineExceeded) {
-				return false, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish updating, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return false, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish updating: %w", hcpClusterName, resourceGroupName, err)
-		}
-
-		switch m := any(operationResult).(type) {
-		case hcpsdk20240610preview.HcpOpenShiftClustersClientUpdateResponse:
-			expect, err := GetHCPCluster20240610(ctx, hcpClient, resourceGroupName, hcpClusterName)
-			if err != nil {
-				if errors.Is(err, context.DeadlineExceeded) {
-					return false, fmt.Errorf("failed getting hcpCluster=%q in resourcegroup=%q, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-				}
-				return false, err
-			}
-			if err := checkOperationResult(&expect.HcpOpenShiftCluster, &m.HcpOpenShiftCluster); err != nil {
-				return false, err
-			}
-			hcpOpenShiftCluster = &m.HcpOpenShiftCluster
-			return true, nil
-		default:
-			return false, fmt.Errorf("unknown type %T", m)
-		}
-	})
-
-	if wait.Interrupted(err) && lastTransientErr != nil {
-		return hcpOpenShiftCluster, fmt.Errorf("update interrupted for hcpCluster=%q in resourcegroup=%q after %d attempts, last transient error: %w, interrupt cause: %w", hcpClusterName, resourceGroupName, attempt, lastTransientErr, err)
-	}
-
-	return hcpOpenShiftCluster, err
 }
 
 // stateConflictBackoff is the retry config for transient state conflicts (ARO-25884).
@@ -411,394 +123,12 @@ func isTransientUpdateError(err error) bool {
 	return false
 }
 
-// UpdateHCPCluster20251223 updates an HCP cluster using the v20251223preview SDK and waits for the operation to complete.
-// Transient 500 and 409 errors are retried automatically with exponential backoff.
-func UpdateHCPCluster20251223(
-	ctx context.Context,
-	hcpClient *hcpsdk20251223preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	update hcpsdk20251223preview.HcpOpenShiftClusterUpdate,
-	timeout time.Duration,
-) (*hcpsdk20251223preview.HcpOpenShiftCluster, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during UpdateHCPCluster20251223 for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	var hcpOpenShiftCluster *hcpsdk20251223preview.HcpOpenShiftCluster
-	var lastTransientErr error
-	attempt := 0
-
-	err := wait.ExponentialBackoffWithContext(ctx, stateConflictBackoff, func(ctx context.Context) (bool, error) {
-		attempt++
-		if attempt > 1 {
-			ginkgo.GinkgoLogr.Info("Retrying cluster update",
-				"cluster", hcpClusterName,
-				"attempt", attempt)
-		}
-
-		poller, err := hcpClient.BeginUpdate(ctx, resourceGroupName, hcpClusterName, update, nil)
-		if err != nil {
-			if isTransientUpdateError(err) {
-				lastTransientErr = err
-				return false, nil
-			}
-			return false, err
-		}
-
-		operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-			Frequency: StandardPollInterval,
-		})
-		if err != nil {
-			if isTransientUpdateError(err) {
-				lastTransientErr = err
-				return false, nil
-			}
-			if errors.Is(err, context.DeadlineExceeded) {
-				return false, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish updating, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return false, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish updating: %w", hcpClusterName, resourceGroupName, err)
-		}
-
-		// The v20251223 variant does not call checkOperationResult because
-		// the v20251223 SDK does not yet have a matching Get response type
-		// to compare against. The v20240610 variant above does this check.
-		hcpOpenShiftCluster = &operationResult.HcpOpenShiftCluster
-		return true, nil
-	})
-
-	if wait.Interrupted(err) && lastTransientErr != nil {
-		return hcpOpenShiftCluster, fmt.Errorf("update interrupted for hcpCluster=%q in resourcegroup=%q after %d attempts, last transient error: %w, interrupt cause: %w", hcpClusterName, resourceGroupName, attempt, lastTransientErr, err)
-	}
-
-	return hcpOpenShiftCluster, err
-}
-
-// GetHCPCluster20240610 fetches an HCP cluster
-func GetHCPCluster20240610(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-) (hcpsdk20240610preview.HcpOpenShiftClustersClientGetResponse, error) {
-	return hcpClient.Get(ctx, resourceGroupName, hcpClusterName, nil)
-}
-
-// DeleteAllHCPClusters20240610 deletes all Clusters within a resource group and waits
-func DeleteAllHCPClusters20240610(
-	ctx context.Context,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	timeout time.Duration,
-) error {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during DeleteAllHCPClusters for resource group %s", timeout.Minutes(), resourceGroupName))
-	defer cancel()
-
-	var hcpClustersWithoutSizeTag []string
-	hcpClusterNames := []string{}
-	hcpClusterPager := hcpClient.NewListByResourceGroupPager(resourceGroupName, nil)
-	for hcpClusterPager.More() {
-		page, err := hcpClusterPager.NextPage(ctx)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return fmt.Errorf("failed listing hcp clusters in resourcegroup=%q, caused by: %w, error: %w", resourceGroupName, context.Cause(ctx), err)
-			}
-			return fmt.Errorf("failed listing hcp clusters in resourcegroup=%q: %w", resourceGroupName, err)
-		}
-		for _, cluster := range page.Value {
-			hcpClusterNames = append(hcpClusterNames, *cluster.Name)
-			if value, set := cluster.Tags[api.TagClusterSizeOverride]; !set || value == nil || *value != string(api.MinimalControlPlanePodSizing) {
-				hcpClustersWithoutSizeTag = append(hcpClustersWithoutSizeTag, *cluster.Name)
-			}
-		}
-	}
-
-	// deletion takes a while, it's worth it to do this in parallel
-	waitGroup, ctx := errgroup.WithContext(ctx)
-	for _, hcpClusterName := range hcpClusterNames {
-		waitGroup.Go(func() error {
-			// prevent a stray panic from exiting the process. Don't do this generally because ginkgo/gomega rely on panics to function.
-			defer utilruntime.HandleCrashWithContext(ctx)
-
-			return DeleteHCPCluster20240610(ctx, hcpClient, resourceGroupName, hcpClusterName, HCPClusterDeletionTimeout)
-		})
-	}
-	if err := waitGroup.Wait(); err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return fmt.Errorf("failed deleting hcp clusters in resourcegroup=%q, caused by: %w, error: %w", resourceGroupName, context.Cause(ctx), err)
-		}
-		// remember that Wait only shows the first error, not all the errors.
-		return fmt.Errorf("at least one hcp cluster failed to delete: %w", err)
-	}
-
-	if len(hcpClustersWithoutSizeTag) > 0 {
-		return &NonConformingClustersError{clusters: hcpClustersWithoutSizeTag}
-	}
-
-	return nil
-}
-
 type NonConformingClustersError struct {
 	clusters []string
 }
 
 func (e *NonConformingClustersError) Error() string {
 	return fmt.Sprintf("the following clusters did not have tags[%s]=%s: %v; we require end-to-end tests to opt into this tag to ensure that the control planes we provision during automated test runs have minimal footprints on our production infrastructure", api.TagClusterSizeOverride, api.MinimalControlPlanePodSizing, e.clusters)
-}
-
-// DeleteNodePool20240610 deletes a nodepool and waits for the operation to complete
-func DeleteNodePool20240610(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20240610preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-	timeout time.Duration,
-) error {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during DeleteNodePool for nodepool %s in cluster %s in resource group %s", timeout.Minutes(), nodePoolName, hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	poller, err := nodePoolsClient.BeginDelete(ctx, resourceGroupName, hcpClusterName, nodePoolName, nil)
-	if err != nil {
-		var respErr *azcore.ResponseError
-		if errors.As(err, &respErr) && respErr.StatusCode == http.StatusConflict {
-			resp, getErr := nodePoolsClient.Get(ctx, resourceGroupName, hcpClusterName, nodePoolName, nil)
-			if getErr == nil && resp.Properties != nil && resp.Properties.ProvisioningState != nil && *resp.Properties.ProvisioningState == hcpsdk20240610preview.ProvisioningStateDeleting {
-				ginkgo.GinkgoLogr.Info("nodepool already deleting, waiting for completion",
-					"nodePool", nodePoolName, "cluster", hcpClusterName, "resourceGroup", resourceGroupName)
-				return waitForNodePoolDeletion20240610(ctx, nodePoolsClient, resourceGroupName, hcpClusterName, nodePoolName)
-			}
-		}
-		if errors.Is(err, context.DeadlineExceeded) {
-			return fmt.Errorf("failed starting nodepool deletion %q for cluster %q in resourcegroup=%q, caused by: %w, error: %w", nodePoolName, hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-		}
-		return err
-	}
-
-	operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return fmt.Errorf("failed waiting for nodepool=%q in cluster=%q resourcegroup=%q to finish deleting, caused by: %w, error: %w", nodePoolName, hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-		}
-		return fmt.Errorf("failed waiting for nodepool=%q in cluster=%q resourcegroup=%q to finish deleting: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.NodePoolsClientDeleteResponse:
-	default:
-		fmt.Printf("#### unknown type %T: content=%v", m, spew.Sdump(m))
-		return fmt.Errorf("unknown type %T", m)
-	}
-
-	return nil
-}
-
-// waitForNodePoolDeletion20240610 polls GET on the nodepool until it returns 404 (deleted).
-func waitForNodePoolDeletion20240610(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20240610preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-) error {
-	for {
-		_, err := nodePoolsClient.Get(ctx, resourceGroupName, hcpClusterName, nodePoolName, nil)
-		if err != nil {
-			var respErr *azcore.ResponseError
-			if errors.As(err, &respErr) && respErr.StatusCode == http.StatusNotFound {
-				ginkgo.GinkgoLogr.Info("nodepool deletion completed",
-					"nodePool", nodePoolName, "cluster", hcpClusterName, "resourceGroup", resourceGroupName)
-				return nil
-			}
-			if errors.Is(err, context.DeadlineExceeded) {
-				return fmt.Errorf("timed out waiting for already-deleting nodepool=%q in cluster=%q resourcegroup=%q to be deleted, caused by: %w, error: %w", nodePoolName, hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return fmt.Errorf("failed polling for deletion of nodepool=%q in cluster=%q resourcegroup=%q: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-		}
-
-		select {
-		case <-ctx.Done():
-			return fmt.Errorf("timed out waiting for already-deleting nodepool=%q in cluster=%q resourcegroup=%q, caused by: %w, error: %w", nodePoolName, hcpClusterName, resourceGroupName, context.Cause(ctx), ctx.Err())
-		case <-time.After(StandardPollInterval):
-		}
-	}
-}
-
-// GetNodePool20240610 fetches a nodepool resource
-func GetNodePool20240610(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20240610preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-) (hcpsdk20240610preview.NodePoolsClientGetResponse, error) {
-	return nodePoolsClient.Get(ctx, resourceGroupName, hcpClusterName, nodePoolName, nil)
-}
-
-// UpdateNodePoolAndWait20240610 sends a PATCH (BeginUpdate) request for a nodepool and waits for completion
-// within the provided timeout. It returns the final update response or an error.
-func UpdateNodePoolAndWait20240610(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20240610preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-	update hcpsdk20240610preview.NodePoolUpdate,
-	timeout time.Duration,
-) (*hcpsdk20240610preview.NodePool, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during UpdateNodePoolAndWait for nodepool %s in cluster %s in resource group %s", timeout.Minutes(), nodePoolName, hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	poller, err := nodePoolsClient.BeginUpdate(ctx, resourceGroupName, hcpClusterName, nodePoolName, update, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to start nodepool %q update in cluster %q resourcegroup=%q: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return nil, fmt.Errorf("failed waiting for nodepool=%q in cluster=%q resourcegroup=%q to finish updating, caused by: %w, error: %w", nodePoolName, hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-		}
-		return nil, fmt.Errorf("failed waiting for nodepool=%q in cluster=%q resourcegroup=%q to finish updating: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.NodePoolsClientUpdateResponse:
-		expect, err := GetNodePool20240610(ctx, nodePoolsClient, resourceGroupName, hcpClusterName, nodePoolName)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed getting nodepool=%q in cluster=%q resourcegroup=%q, caused by: %w, error: %w", nodePoolName, hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return nil, err
-		}
-		err = checkOperationResult(&expect.NodePool, &m.NodePool)
-		if err != nil {
-			return nil, err
-		}
-		return &m.NodePool, nil
-	default:
-		return nil, fmt.Errorf("unknown type %T", m)
-	}
-}
-
-// CreateOrUpdateExternalAuthAndWait20240610 creates or updates an external auth on an HCP cluster and waits
-func CreateOrUpdateExternalAuthAndWait20240610(
-	ctx context.Context,
-	externalAuthClient *hcpsdk20240610preview.ExternalAuthsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	externalAuthName string,
-	externalAuth hcpsdk20240610preview.ExternalAuth,
-	timeout time.Duration,
-) (*hcpsdk20240610preview.ExternalAuth, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during CreateOrUpdateExternalAuthAndWait for external auth %s in cluster %s in resource group %s", timeout.Minutes(), externalAuthName, hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	pollerResp, err := externalAuthClient.BeginCreateOrUpdate(
-		ctx,
-		resourceGroupName,
-		hcpClusterName,
-		externalAuthName,
-		externalAuth,
-		nil,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed creating external auth %q in resourcegroup=%q for cluster=%q: %w", externalAuthName, resourceGroupName, hcpClusterName, err)
-	}
-	operationResult, err := pollerResp.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return nil, fmt.Errorf("failed waiting for external auth %q in resourcegroup=%q for cluster=%q to finish creating or updating, caused by: %w, error: %w", externalAuthName, resourceGroupName, hcpClusterName, context.Cause(ctx), err)
-		}
-		return nil, fmt.Errorf("failed waiting for external auth %q in resourcegroup=%q for cluster=%q to finish creating or updating: %w", externalAuthName, resourceGroupName, hcpClusterName, err)
-	}
-
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.ExternalAuthsClientCreateOrUpdateResponse:
-		// Verify the operationResult content matches the current external auth model.
-		// When an asynchronous operation completes successfully, the RP's result
-		// endpoint for the operation is supposed to respond as though the operation
-		// were completed synchronously. In production, ARM would call this endpoint
-		// automatically. In this context, the poller calls it automatically.
-		expect, err := GetExternalAuth20240610(ctx, externalAuthClient, resourceGroupName, hcpClusterName, externalAuthName)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed getting external auth %q in resourcegroup=%q for cluster=%q, caused by: %w, error: %w", externalAuthName, resourceGroupName, hcpClusterName, context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed getting external auth %q in resourcegroup=%q for cluster=%q: %w", externalAuthName, resourceGroupName, hcpClusterName, err)
-		}
-		err = checkOperationResult(&expect.ExternalAuth, &m.ExternalAuth)
-		if err != nil {
-			return nil, err
-		}
-		return &m.ExternalAuth, nil
-	default:
-		fmt.Printf("#### unknown type %T: content=%v", m, spew.Sdump(m))
-		return nil, fmt.Errorf("unknown type %T", m)
-	}
-}
-
-// CreateExternalAuthAndWait creates a an external auth on an HCP cluster and waits
-func GetExternalAuth20240610(
-	ctx context.Context,
-	externalAuthClient *hcpsdk20240610preview.ExternalAuthsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	externalAuthName string,
-) (hcpsdk20240610preview.ExternalAuthsClientGetResponse, error) {
-	return externalAuthClient.Get(
-		ctx,
-		resourceGroupName,
-		hcpClusterName,
-		externalAuthName,
-		&hcpsdk20240610preview.ExternalAuthsClientGetOptions{},
-	)
-}
-
-// DeleteExternalAuthAndWait20240610 deletes a an external auth on an HCP cluster and waits
-func DeleteExternalAuthAndWait20240610(
-	ctx context.Context,
-	externalAuthClient *hcpsdk20240610preview.ExternalAuthsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	externalAuthName string,
-	timeout time.Duration,
-) error {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during DeleteExternalAuthAndWait for external auth %s in cluster %s in resource group %s", timeout.Minutes(), externalAuthName, hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	pollerResp, err := externalAuthClient.BeginDelete(
-		ctx,
-		resourceGroupName,
-		hcpClusterName,
-		externalAuthName,
-		nil,
-	)
-	if err != nil {
-		return fmt.Errorf("failed deleting external auth %q in resourcegroup=%q for cluster=%q: %w", externalAuthName, resourceGroupName, hcpClusterName, err)
-	}
-	operationResult, err := pollerResp.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return fmt.Errorf("failed waiting for external auth %q in resourcegroup=%q for cluster=%q to finish deleting, caused by: %w, error: %w", externalAuthName, resourceGroupName, hcpClusterName, context.Cause(ctx), err)
-		}
-		return fmt.Errorf("failed waiting for external auth %q in resourcegroup=%q for cluster=%q to finish deleting: %w", externalAuthName, resourceGroupName, hcpClusterName, err)
-	}
-
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.ExternalAuthsClientDeleteResponse:
-		return nil
-	default:
-		fmt.Printf("#### unknown type %T: content=%v", m, spew.Sdump(m))
-		return fmt.Errorf("unknown type %T", m)
-	}
 }
 
 func CreateClusterRoleBinding(ctx context.Context, subject string, adminRESTConfig *rest.Config) error {
@@ -860,271 +190,6 @@ func CreateTestDockerConfigSecret(host, username, password, email, secretName, n
 			corev1.DockerConfigJsonKey: dockerConfigJSON,
 		},
 	}, nil
-}
-
-func BeginCreateHCPCluster20240610(
-	ctx context.Context,
-	logger logr.Logger,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	clusterParams ClusterParams20240610,
-	location string,
-) (*runtime.Poller[hcpsdk20240610preview.HcpOpenShiftClustersClientCreateOrUpdateResponse], error) {
-	cluster := BuildHCPClusterFromParams20240610(clusterParams, location)
-	logger.Info("Starting HCP cluster creation", "clusterName", hcpClusterName, "resourceGroup", resourceGroupName)
-	poller, err := hcpClient.BeginCreateOrUpdate(ctx, resourceGroupName, hcpClusterName, cluster, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed starting cluster creation %q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-	}
-	return poller, nil
-}
-
-// CreateHCPClusterAndWait20240610 Note that the timeout parameter will only take effect if its value is greater than 0. Otherwise,
-// the function won't wait for the deployment to be ready.
-func CreateHCPClusterAndWait20240610(
-	ctx context.Context,
-	logger logr.Logger,
-	hcpClient *hcpsdk20240610preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	cluster hcpsdk20240610preview.HcpOpenShiftCluster,
-	timeout time.Duration,
-) (*hcpsdk20240610preview.HcpOpenShiftCluster, error) {
-	if timeout > 0*time.Second {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during CreateHCPClusterAndWait for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-		defer cancel()
-	}
-
-	logger.Info("Starting HCP cluster creation", "clusterName", hcpClusterName, "resourceGroup", resourceGroupName, "version", cluster.Properties.Version.ID, "channelGroup", cluster.Properties.Version.ChannelGroup)
-	poller, err := hcpClient.BeginCreateOrUpdate(ctx, resourceGroupName, hcpClusterName, cluster, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed starting cluster creation %q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-	}
-
-	if timeout > 0*time.Second {
-		operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-			Frequency: StandardPollInterval,
-		})
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed waiting for cluster=%q in resourcegroup=%q to finish creating, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed waiting for cluster=%q in resourcegroup=%q to finish creating: %w", hcpClusterName, resourceGroupName, err)
-		}
-		switch m := any(operationResult).(type) {
-		case hcpsdk20240610preview.HcpOpenShiftClustersClientCreateOrUpdateResponse:
-			// Verify the operationResult content matches the current cluster model.
-			// When an asynchronous operation completes successfully, the RP's result
-			// endpoint for the operation is supposed to respond as though the operation
-			// were completed synchronously. In production, ARM would call this endpoint
-			// automatically. In this context, the poller calls it automatically.
-			expect, err := GetHCPCluster20240610(ctx, hcpClient, resourceGroupName, hcpClusterName)
-			if err != nil {
-				if errors.Is(err, context.DeadlineExceeded) {
-					return nil, fmt.Errorf("failed getting cluster=%q in resourcegroup=%q, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-				}
-				return nil, err
-			}
-			err = checkOperationResult(&expect.HcpOpenShiftCluster, &m.HcpOpenShiftCluster)
-			if err != nil {
-				return nil, err
-			}
-			return &m.HcpOpenShiftCluster, nil
-		default:
-			fmt.Printf("unknown type %T: content=%v", m, spew.Sdump(m))
-			return nil, fmt.Errorf("unknown type %T", m)
-		}
-	} else {
-		_, err := poller.Poll(ctx)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed checking for deployment %q in resourcegroup=%q, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed checking for deployment %q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-		}
-		return nil, nil
-	}
-
-}
-
-func BuildHCPClusterFromParams20240610(
-	parameters ClusterParams20240610,
-	location string,
-) hcpsdk20240610preview.HcpOpenShiftCluster {
-
-	return hcpsdk20240610preview.HcpOpenShiftCluster{
-		Location: to.Ptr(location),
-		Identity: parameters.Identity,
-		Tags:     parameters.Tags,
-		Properties: &hcpsdk20240610preview.HcpOpenShiftClusterProperties{
-			Version: &hcpsdk20240610preview.VersionProfile{
-				ID:           to.Ptr(parameters.OpenshiftVersionId),
-				ChannelGroup: to.Ptr(parameters.ChannelGroup),
-			},
-			Platform: &hcpsdk20240610preview.PlatformProfile{
-				ManagedResourceGroup:   to.Ptr(parameters.ManagedResourceGroupName),
-				NetworkSecurityGroupID: to.Ptr(parameters.NsgResourceID),
-				SubnetID:               to.Ptr(parameters.SubnetResourceID),
-				OperatorsAuthentication: &hcpsdk20240610preview.OperatorsAuthenticationProfile{
-					UserAssignedIdentities: parameters.UserAssignedIdentitiesProfile,
-				}},
-			Network: &hcpsdk20240610preview.NetworkProfile{
-				NetworkType: to.Ptr(hcpsdk20240610preview.NetworkType(parameters.Network.NetworkType)),
-				PodCIDR:     to.Ptr(parameters.Network.PodCIDR),
-				ServiceCIDR: to.Ptr(parameters.Network.ServiceCIDR),
-				MachineCIDR: to.Ptr(parameters.Network.MachineCIDR),
-				HostPrefix:  to.Ptr(parameters.Network.HostPrefix),
-			},
-			API: &hcpsdk20240610preview.APIProfile{
-				Visibility:      to.Ptr(hcpsdk20240610preview.Visibility(parameters.APIVisibility)),
-				AuthorizedCIDRs: parameters.AuthorizedCIDRs,
-			},
-			ClusterImageRegistry: &hcpsdk20240610preview.ClusterImageRegistryProfile{
-				State: to.Ptr(hcpsdk20240610preview.ClusterImageRegistryState(parameters.ImageRegistryState)),
-			},
-			Etcd: &hcpsdk20240610preview.EtcdProfile{
-				DataEncryption: &hcpsdk20240610preview.EtcdDataEncryptionProfile{
-					KeyManagementMode: to.Ptr(hcpsdk20240610preview.EtcdDataEncryptionKeyManagementModeType(parameters.EncryptionKeyManagementMode)),
-					CustomerManaged: &hcpsdk20240610preview.CustomerManagedEncryptionProfile{
-						EncryptionType: to.Ptr(hcpsdk20240610preview.CustomerManagedEncryptionType(parameters.EncryptionType)),
-						Kms: &hcpsdk20240610preview.KmsEncryptionProfile{
-							ActiveKey: &hcpsdk20240610preview.KmsKey{
-								VaultName: to.Ptr(parameters.KeyVaultName),
-								Name:      to.Ptr(parameters.EtcdEncryptionKeyName),
-								Version:   to.Ptr(parameters.EtcdEncryptionKeyVersion),
-							},
-						},
-					},
-				},
-			},
-			Autoscaling: parameters.Autoscaling,
-		},
-	}
-}
-
-func CreateNodePoolAndWait20240610(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20240610preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-	nodePool hcpsdk20240610preview.NodePool,
-	timeout time.Duration,
-) (*hcpsdk20240610preview.NodePool, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during CreateNodePoolAndWait for nodepool %s in cluster %s in resource group %s", timeout.Minutes(), nodePoolName, hcpClusterName, resourceGroupName))
-	defer cancel()
-	poller, err := nodePoolsClient.BeginCreateOrUpdate(ctx, resourceGroupName, hcpClusterName, nodePoolName, nodePool, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed starting nodepool creation %q for cluster %q in resourcegroup=%q: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed waiting for nodepool=%q for cluster %q in resourcegroup=%q to finish creating: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-	switch m := any(operationResult).(type) {
-	case hcpsdk20240610preview.NodePoolsClientCreateOrUpdateResponse:
-		// Verify the operationResult content matches the current node pool model.
-		// When an asynchronous operation completes successfully, the RP's result
-		// endpoint for the operation is supposed to respond as though the operation
-		// were completed synchronously. In production, ARM would call this endpoint
-		// automatically. In this context, the poller calls it automatically.
-		expect, err := GetNodePool20240610(ctx, nodePoolsClient, resourceGroupName, hcpClusterName, nodePoolName)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed to get nodepool, caused by: %w, error: %w", context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed to get nodepool %q in cluster %q resourcegroup=%q after creation: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-		}
-		err = checkOperationResult(&expect.NodePool, &m.NodePool)
-		if err != nil {
-			return nil, err
-		}
-		return &m.NodePool, nil
-	default:
-		fmt.Printf("unknown type %T: content=%v", m, spew.Sdump(m))
-		return nil, fmt.Errorf("unknown type %T", m)
-	}
-}
-
-func BuildNodePoolFromParams20240610(
-	parameters NodePoolParams20240610,
-	location string,
-) hcpsdk20240610preview.NodePool {
-
-	nodePool := hcpsdk20240610preview.NodePool{
-		Location: to.Ptr(location),
-		Tags:     parameters.Tags,
-		Properties: &hcpsdk20240610preview.NodePoolProperties{
-			Version: &hcpsdk20240610preview.NodePoolVersionProfile{
-				ID:           to.Ptr(parameters.OpenshiftVersionId),
-				ChannelGroup: to.Ptr(parameters.ChannelGroup),
-			},
-			NodeDrainTimeoutMinutes: parameters.NodeDrainTimeoutMinutes,
-			Platform: &hcpsdk20240610preview.NodePoolPlatformProfile{
-				VMSize: to.Ptr(parameters.VMSize),
-				OSDisk: &hcpsdk20240610preview.OsDiskProfile{
-					SizeGiB:                to.Ptr(parameters.OSDiskSizeGiB),
-					DiskStorageAccountType: to.Ptr(hcpsdk20240610preview.DiskStorageAccountType(parameters.DiskStorageAccountType)),
-				},
-				AvailabilityZone: to.Ptr(parameters.AvailabilityZone),
-			},
-		},
-	}
-
-	if parameters.AutoScaling != nil {
-		nodePool.Properties.AutoScaling = &hcpsdk20240610preview.NodePoolAutoScaling{
-			Min: to.Ptr(parameters.AutoScaling.Min),
-			Max: to.Ptr(parameters.AutoScaling.Max),
-		}
-	} else {
-		nodePool.Properties.Replicas = to.Ptr(parameters.Replicas)
-	}
-
-	return nodePool
-}
-
-func BuildNodePoolFromParams20251223(
-	parameters NodePoolParams20251223,
-	location string,
-) hcpsdk20251223preview.NodePool {
-
-	nodePool := hcpsdk20251223preview.NodePool{
-		Location: to.Ptr(location),
-		Tags:     parameters.Tags,
-		Properties: &hcpsdk20251223preview.NodePoolProperties{
-			Version: &hcpsdk20251223preview.NodePoolVersionProfile{
-				ID:           to.Ptr(parameters.OpenshiftVersionId),
-				ChannelGroup: to.Ptr(parameters.ChannelGroup),
-			},
-			NodeDrainTimeoutMinutes: parameters.NodeDrainTimeoutMinutes,
-			Platform: &hcpsdk20251223preview.NodePoolPlatformProfile{
-				VMSize: to.Ptr(parameters.VMSize),
-				OSDisk: &hcpsdk20251223preview.OsDiskProfile{
-					SizeGiB:                to.Ptr(parameters.OSDiskSizeGiB),
-					DiskStorageAccountType: to.Ptr(hcpsdk20251223preview.DiskStorageAccountType(parameters.DiskStorageAccountType)),
-					DiskType:               to.Ptr(parameters.DiskType),
-				},
-				AvailabilityZone: to.Ptr(parameters.AvailabilityZone),
-			},
-			AutoRepair: to.Ptr(parameters.AutoRepair),
-		},
-	}
-
-	if parameters.AutoScaling != nil {
-		nodePool.Properties.AutoScaling = &hcpsdk20251223preview.NodePoolAutoScaling{
-			Min: to.Ptr(parameters.AutoScaling.Min),
-			Max: to.Ptr(parameters.AutoScaling.Max),
-		}
-	} else {
-		nodePool.Properties.Replicas = to.Ptr(parameters.Replicas)
-	}
-
-	return nodePool
 }
 
 // Helper to generate SSH key pair
@@ -1219,178 +284,6 @@ func convertViaJSON[T any](src any) (*T, error) {
 	return &dst, nil
 }
 
-func BuildHCPClusterFromParams20251223(
-	parameters ClusterParams20251223,
-	location string,
-	imageDigestMirrors []*hcpsdk20251223preview.ImageDigestMirror,
-) (hcpsdk20251223preview.HcpOpenShiftCluster, error) {
-	// Convert identity types from old SDK to new SDK via JSON round-trip
-	var identity *hcpsdk20251223preview.ManagedServiceIdentity
-	if parameters.Identity != nil {
-		var err error
-		identity, err = convertViaJSON[hcpsdk20251223preview.ManagedServiceIdentity](parameters.Identity)
-		if err != nil {
-			return hcpsdk20251223preview.HcpOpenShiftCluster{}, fmt.Errorf("failed to convert Identity: %w", err)
-		}
-	}
-
-	var uamis *hcpsdk20251223preview.UserAssignedIdentitiesProfile
-	if parameters.UserAssignedIdentitiesProfile != nil {
-		var err error
-		uamis, err = convertViaJSON[hcpsdk20251223preview.UserAssignedIdentitiesProfile](parameters.UserAssignedIdentitiesProfile)
-		if err != nil {
-			return hcpsdk20251223preview.HcpOpenShiftCluster{}, fmt.Errorf("failed to convert UserAssignedIdentitiesProfile: %w", err)
-		}
-	}
-
-	return hcpsdk20251223preview.HcpOpenShiftCluster{
-		Location: to.Ptr(location),
-		Identity: identity,
-		Tags:     parameters.Tags,
-		Properties: &hcpsdk20251223preview.HcpOpenShiftClusterProperties{
-			Version: &hcpsdk20251223preview.VersionProfile{
-				ID:           to.Ptr(parameters.OpenshiftVersionId),
-				ChannelGroup: to.Ptr(parameters.ChannelGroup),
-			},
-			Platform: &hcpsdk20251223preview.PlatformProfile{
-				ManagedResourceGroup:    to.Ptr(parameters.ManagedResourceGroupName),
-				NetworkSecurityGroupID:  to.Ptr(parameters.NsgResourceID),
-				SubnetID:                to.Ptr(parameters.SubnetResourceID),
-				VnetIntegrationSubnetID: to.Ptr(parameters.VnetIntegrationSubnetID),
-				OperatorsAuthentication: &hcpsdk20251223preview.OperatorsAuthenticationProfile{
-					UserAssignedIdentities: uamis,
-				},
-			},
-			Network: &hcpsdk20251223preview.NetworkProfile{
-				NetworkType: to.Ptr(hcpsdk20251223preview.NetworkType(parameters.Network.NetworkType)),
-				PodCIDR:     to.Ptr(parameters.Network.PodCIDR),
-				ServiceCIDR: to.Ptr(parameters.Network.ServiceCIDR),
-				MachineCIDR: to.Ptr(parameters.Network.MachineCIDR),
-				HostPrefix:  to.Ptr(parameters.Network.HostPrefix),
-			},
-			API: &hcpsdk20251223preview.APIProfile{
-				Visibility:      to.Ptr(hcpsdk20251223preview.Visibility(parameters.APIVisibility)),
-				AuthorizedCIDRs: parameters.AuthorizedCIDRs,
-			},
-			ClusterImageRegistry: &hcpsdk20251223preview.ClusterImageRegistryProfile{
-				State: to.Ptr(hcpsdk20251223preview.ClusterImageRegistryState(parameters.ImageRegistryState)),
-			},
-			Etcd: &hcpsdk20251223preview.EtcdProfile{
-				DataEncryption: &hcpsdk20251223preview.EtcdDataEncryptionProfile{
-					KeyManagementMode: to.Ptr(hcpsdk20251223preview.EtcdDataEncryptionKeyManagementModeType(parameters.EncryptionKeyManagementMode)),
-					CustomerManaged: &hcpsdk20251223preview.CustomerManagedEncryptionProfile{
-						EncryptionType: to.Ptr(hcpsdk20251223preview.CustomerManagedEncryptionType(parameters.EncryptionType)),
-						Kms: &hcpsdk20251223preview.KmsEncryptionProfile{
-							VaultName:  to.Ptr(parameters.KeyVaultName),
-							Visibility: to.Ptr(hcpsdk20251223preview.KeyVaultVisibility(parameters.KeyVaultVisibility)),
-							ActiveKey: &hcpsdk20251223preview.KmsKey{
-								Name:    to.Ptr(parameters.EtcdEncryptionKeyName),
-								Version: to.Ptr(parameters.EtcdEncryptionKeyVersion),
-							},
-						},
-					},
-				},
-			},
-			ImageDigestMirrors: imageDigestMirrors,
-		},
-	}, nil
-}
-
-// CreateHCPClusterAndWait20251223 creates an HCP cluster using the v20251223preview API and waits for completion.
-func CreateHCPClusterAndWait20251223(
-	ctx context.Context,
-	logger logr.Logger,
-	hcpClient *hcpsdk20251223preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	cluster hcpsdk20251223preview.HcpOpenShiftCluster,
-	timeout time.Duration,
-) (*hcpsdk20251223preview.HcpOpenShiftCluster, error) {
-	if timeout > 0*time.Second {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during CreateHCPCluster20251223AndWait for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-		defer cancel()
-	}
-
-	logger.Info("Starting HCP cluster creation (v20251223preview)", "clusterName", hcpClusterName, "resourceGroup", resourceGroupName)
-	poller, err := hcpClient.BeginCreateOrUpdate(ctx, resourceGroupName, hcpClusterName, cluster, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed starting cluster creation %q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-	}
-
-	if timeout > 0*time.Second {
-		operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-			Frequency: StandardPollInterval,
-		})
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed waiting for cluster=%q in resourcegroup=%q to finish creating, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed waiting for cluster=%q in resourcegroup=%q to finish creating: %w", hcpClusterName, resourceGroupName, err)
-		}
-		switch m := any(operationResult).(type) {
-		case hcpsdk20251223preview.HcpOpenShiftClustersClientCreateOrUpdateResponse:
-			return &m.HcpOpenShiftCluster, nil
-		default:
-			fmt.Printf("unknown type %T: content=%v", m, spew.Sdump(m))
-			return nil, fmt.Errorf("unknown type %T", m)
-		}
-	} else {
-		_, err := poller.Poll(ctx)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed checking for deployment %q in resourcegroup=%q, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed checking for deployment %q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-		}
-		return nil, nil
-	}
-}
-
-// Verifies that a nodepool created using framework has DiskStorageAccountType set to the framework default "StandardSSD_LRS"
-func ValidateNodePoolDiskStorageAccountType20240610(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20240610preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-) error {
-	nodePoolResp, err := GetNodePool20240610(ctx, nodePoolsClient, resourceGroupName, hcpClusterName, nodePoolName)
-	if err != nil {
-		return fmt.Errorf("failed to get nodepool %s: %w", nodePoolName, err)
-	}
-
-	nodePool := nodePoolResp.NodePool
-
-	// Verify the nodepool exists and has the expected structure
-	if nodePool.Properties == nil {
-		return fmt.Errorf("nodepool %s has no properties", nodePoolName)
-	}
-
-	if nodePool.Properties.Platform == nil {
-		return fmt.Errorf("nodepool %s has no platform configuration", nodePoolName)
-	}
-
-	if nodePool.Properties.Platform.OSDisk == nil {
-		return fmt.Errorf("nodepool %s has no OS disk configuration", nodePoolName)
-	}
-
-	if nodePool.Properties.Platform.OSDisk.DiskStorageAccountType == nil {
-		return fmt.Errorf("nodepool %s has no DiskStorageAccountType set", nodePoolName)
-	}
-
-	// Verify the framework default (StandardSSD_LRS) overrode the API default (Premium_LRS)
-	expectedDiskType := "StandardSSD_LRS"
-	actualDiskType := string(*nodePool.Properties.Platform.OSDisk.DiskStorageAccountType)
-
-	if actualDiskType != expectedDiskType {
-		return fmt.Errorf("nodepool %s has incorrect DiskStorageAccountType: expected %s (framework default), got %s",
-			nodePoolName, expectedDiskType, actualDiskType)
-	}
-
-	return nil
-}
-
 func HasNodeLabel(nodes []corev1.Node, key, value string, expectedCount ...int) bool {
 	count := 0
 	for _, node := range nodes {
@@ -1423,346 +316,3 @@ func HasNodeTaint(nodes []corev1.Node, key, value string, effect corev1.TaintEff
 
 	return count == expectedCount[0]
 }
-
-// CreateNodePoolAndWait20251223 creates a nodepool using the v20251223preview API and waits for completion.
-func CreateNodePoolAndWait20251223(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20251223preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-	nodePool hcpsdk20251223preview.NodePool,
-	timeout time.Duration,
-) (*hcpsdk20251223preview.NodePool, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during CreateNodePoolAndWait20251223 for nodepool %s in cluster %s in resource group %s", timeout.Minutes(), nodePoolName, hcpClusterName, resourceGroupName))
-	defer cancel()
-	poller, err := nodePoolsClient.BeginCreateOrUpdate(ctx, resourceGroupName, hcpClusterName, nodePoolName, nodePool, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed starting nodepool creation %q for cluster %q in resourcegroup=%q: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed waiting for nodepool=%q for cluster %q in resourcegroup=%q to finish creating: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	// Verify the LRO result body matches a fresh GET, per ARM LRO contract.
-	expect, err := GetNodePool20251223(ctx, nodePoolsClient, resourceGroupName, hcpClusterName, nodePoolName)
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return nil, fmt.Errorf("failed to get nodepool, caused by: %w, error: %w", context.Cause(ctx), err)
-		}
-		return nil, err
-	}
-	if err := checkOperationResult(expect, &operationResult.NodePool); err != nil {
-		return nil, err
-	}
-
-	return &operationResult.NodePool, nil
-}
-
-// GetNodePool20251223 retrieves a nodepool using the v20251223preview API.
-func GetNodePool20251223(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20251223preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-) (*hcpsdk20251223preview.NodePool, error) {
-	resp, err := nodePoolsClient.Get(ctx, resourceGroupName, hcpClusterName, nodePoolName, nil)
-	if err != nil {
-		return nil, err
-	}
-	return &resp.NodePool, nil
-}
-
-func BuildHCPClusterFromParams20260630(
-	parameters ClusterParams20260630,
-	location string,
-	imageDigestMirrors []*hcpsdk20260630preview.ImageDigestMirror,
-) (hcpsdk20260630preview.HcpOpenShiftCluster, error) {
-	// Convert identity types from old SDK to new SDK via JSON round-trip
-	var identity *hcpsdk20260630preview.ManagedServiceIdentity
-	if parameters.Identity != nil {
-		var err error
-		identity, err = convertViaJSON[hcpsdk20260630preview.ManagedServiceIdentity](parameters.Identity)
-		if err != nil {
-			return hcpsdk20260630preview.HcpOpenShiftCluster{}, fmt.Errorf("failed to convert Identity: %w", err)
-		}
-	}
-
-	var uamis *hcpsdk20260630preview.UserAssignedIdentitiesProfile
-	if parameters.UserAssignedIdentitiesProfile != nil {
-		var err error
-		uamis, err = convertViaJSON[hcpsdk20260630preview.UserAssignedIdentitiesProfile](parameters.UserAssignedIdentitiesProfile)
-		if err != nil {
-			return hcpsdk20260630preview.HcpOpenShiftCluster{}, fmt.Errorf("failed to convert UserAssignedIdentitiesProfile: %w", err)
-		}
-	}
-
-	return hcpsdk20260630preview.HcpOpenShiftCluster{
-		Location: to.Ptr(location),
-		Identity: identity,
-		Tags:     parameters.Tags,
-		Properties: &hcpsdk20260630preview.HcpOpenShiftClusterProperties{
-			Version: &hcpsdk20260630preview.VersionProfile{
-				ID:           to.Ptr(parameters.OpenshiftVersionId),
-				ChannelGroup: to.Ptr(parameters.ChannelGroup),
-			},
-			Platform: &hcpsdk20260630preview.PlatformProfile{
-				ManagedResourceGroup:    to.Ptr(parameters.ManagedResourceGroupName),
-				NetworkSecurityGroupID:  to.Ptr(parameters.NsgResourceID),
-				SubnetID:                to.Ptr(parameters.SubnetResourceID),
-				VnetIntegrationSubnetID: to.Ptr(parameters.VnetIntegrationSubnetID),
-				OperatorsAuthentication: &hcpsdk20260630preview.OperatorsAuthenticationProfile{
-					UserAssignedIdentities: uamis,
-				},
-			},
-			Network: &hcpsdk20260630preview.NetworkProfile{
-				NetworkType: to.Ptr(hcpsdk20260630preview.NetworkType(parameters.Network.NetworkType)),
-				PodCIDR:     to.Ptr(parameters.Network.PodCIDR),
-				ServiceCIDR: to.Ptr(parameters.Network.ServiceCIDR),
-				MachineCIDR: to.Ptr(parameters.Network.MachineCIDR),
-				HostPrefix:  to.Ptr(parameters.Network.HostPrefix),
-			},
-			API: &hcpsdk20260630preview.APIProfile{
-				Visibility:      to.Ptr(hcpsdk20260630preview.Visibility(parameters.APIVisibility)),
-				AuthorizedCIDRs: parameters.AuthorizedCIDRs,
-			},
-			Ingress: &hcpsdk20260630preview.IngressProfile{
-				Type: to.Ptr(hcpsdk20260630preview.IngressType(parameters.IngressType)),
-			},
-			ClusterImageRegistry: &hcpsdk20260630preview.ClusterImageRegistryProfile{
-				State: to.Ptr(hcpsdk20260630preview.ClusterImageRegistryState(parameters.ImageRegistryState)),
-			},
-			CryptoRestrictions: parameters.CryptoRestrictions,
-			Etcd: &hcpsdk20260630preview.EtcdProfile{
-				DataEncryption: &hcpsdk20260630preview.EtcdDataEncryptionProfile{
-					KeyManagementMode: to.Ptr(hcpsdk20260630preview.EtcdDataEncryptionKeyManagementModeType(parameters.EncryptionKeyManagementMode)),
-					CustomerManaged: &hcpsdk20260630preview.CustomerManagedEncryptionProfile{
-						EncryptionType: to.Ptr(hcpsdk20260630preview.CustomerManagedEncryptionType(parameters.EncryptionType)),
-						Kms: &hcpsdk20260630preview.KmsEncryptionProfile{
-							VaultName:  to.Ptr(parameters.KeyVaultName),
-							Visibility: to.Ptr(hcpsdk20260630preview.KeyVaultVisibility(parameters.KeyVaultVisibility)),
-							ActiveKey: &hcpsdk20260630preview.KmsKey{
-								Name:    to.Ptr(parameters.EtcdEncryptionKeyName),
-								Version: to.Ptr(parameters.EtcdEncryptionKeyVersion),
-							},
-						},
-					},
-				},
-			},
-			ImageDigestMirrors: imageDigestMirrors,
-		},
-	}, nil
-}
-
-// CreateHCPClusterAndWait20260630 creates an HCP cluster using the v20260630preview API and waits for completion.
-func CreateHCPClusterAndWait20260630(
-	ctx context.Context,
-	logger logr.Logger,
-	hcpClient *hcpsdk20260630preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	cluster hcpsdk20260630preview.HcpOpenShiftCluster,
-	timeout time.Duration,
-) (*hcpsdk20260630preview.HcpOpenShiftCluster, error) {
-	if timeout > 0*time.Second {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during CreateHCPCluster20260630AndWait for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-		defer cancel()
-	}
-
-	logger.Info("Starting HCP cluster creation (v20260630preview)", "clusterName", hcpClusterName, "resourceGroup", resourceGroupName)
-	poller, err := hcpClient.BeginCreateOrUpdate(ctx, resourceGroupName, hcpClusterName, cluster, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed starting cluster creation %q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-	}
-
-	if timeout > 0*time.Second {
-		operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-			Frequency: StandardPollInterval,
-		})
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed waiting for cluster=%q in resourcegroup=%q to finish creating, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed waiting for cluster=%q in resourcegroup=%q to finish creating: %w", hcpClusterName, resourceGroupName, err)
-		}
-		switch m := any(operationResult).(type) {
-		case hcpsdk20260630preview.HcpOpenShiftClustersClientCreateOrUpdateResponse:
-			return &m.HcpOpenShiftCluster, nil
-		default:
-			fmt.Printf("unknown type %T: content=%v", m, spew.Sdump(m))
-			return nil, fmt.Errorf("unknown type %T", m)
-		}
-	} else {
-		_, err := poller.Poll(ctx)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				return nil, fmt.Errorf("failed checking for deployment %q in resourcegroup=%q, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return nil, fmt.Errorf("failed checking for deployment %q in resourcegroup=%q: %w", hcpClusterName, resourceGroupName, err)
-		}
-		return nil, nil
-	}
-}
-
-// UpdateHCPCluster20260630 updates an HCP cluster using the v20260630preview SDK and waits for the operation to complete.
-// Transient 500 and 409 errors are retried automatically with exponential backoff.
-func UpdateHCPCluster20260630(
-	ctx context.Context,
-	hcpClient *hcpsdk20260630preview.HcpOpenShiftClustersClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	update hcpsdk20260630preview.HcpOpenShiftClusterUpdate,
-	timeout time.Duration,
-) (*hcpsdk20260630preview.HcpOpenShiftCluster, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during UpdateHCPCluster20260630 for cluster %s in resource group %s", timeout.Minutes(), hcpClusterName, resourceGroupName))
-	defer cancel()
-
-	var hcpOpenShiftCluster *hcpsdk20260630preview.HcpOpenShiftCluster
-	var lastTransientErr error
-	attempt := 0
-
-	err := wait.ExponentialBackoffWithContext(ctx, stateConflictBackoff, func(ctx context.Context) (bool, error) {
-		attempt++
-		if attempt > 1 {
-			ginkgo.GinkgoLogr.Info("Retrying cluster update",
-				"cluster", hcpClusterName,
-				"attempt", attempt)
-		}
-
-		poller, err := hcpClient.BeginUpdate(ctx, resourceGroupName, hcpClusterName, update, nil)
-		if err != nil {
-			if isTransientUpdateError(err) {
-				lastTransientErr = err
-				return false, nil
-			}
-			return false, err
-		}
-
-		operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-			Frequency: StandardPollInterval,
-		})
-		if err != nil {
-			if isTransientUpdateError(err) {
-				lastTransientErr = err
-				return false, nil
-			}
-			if errors.Is(err, context.DeadlineExceeded) {
-				return false, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish updating, caused by: %w, error: %w", hcpClusterName, resourceGroupName, context.Cause(ctx), err)
-			}
-			return false, fmt.Errorf("failed waiting for hcpCluster=%q in resourcegroup=%q to finish updating: %w", hcpClusterName, resourceGroupName, err)
-		}
-
-		// The v20260630 variant does not call checkOperationResult because
-		// the v20260630 SDK does not yet have a matching Get response type
-		// to compare against. The v20240610 variant above does this check.
-		hcpOpenShiftCluster = &operationResult.HcpOpenShiftCluster
-		return true, nil
-	})
-
-	if wait.Interrupted(err) && lastTransientErr != nil {
-		return hcpOpenShiftCluster, fmt.Errorf("update interrupted for hcpCluster=%q in resourcegroup=%q after %d attempts, last transient error: %w, interrupt cause: %w", hcpClusterName, resourceGroupName, attempt, lastTransientErr, err)
-	}
-
-	return hcpOpenShiftCluster, err
-}
-
-func BuildNodePoolFromParams20260630(
-	parameters NodePoolParams20260630,
-	location string,
-) hcpsdk20260630preview.NodePool {
-
-	nodePool := hcpsdk20260630preview.NodePool{
-		Location: to.Ptr(location),
-		Tags:     parameters.Tags,
-		Properties: &hcpsdk20260630preview.NodePoolProperties{
-			Version: &hcpsdk20260630preview.NodePoolVersionProfile{
-				ID:           to.Ptr(parameters.OpenshiftVersionId),
-				ChannelGroup: to.Ptr(parameters.ChannelGroup),
-			},
-			NodeDrainTimeoutMinutes: parameters.NodeDrainTimeoutMinutes,
-			Platform: &hcpsdk20260630preview.NodePoolPlatformProfile{
-				VMSize: to.Ptr(parameters.VMSize),
-				OSDisk: &hcpsdk20260630preview.OsDiskProfile{
-					SizeGiB:                to.Ptr(parameters.OSDiskSizeGiB),
-					DiskStorageAccountType: to.Ptr(hcpsdk20260630preview.DiskStorageAccountType(parameters.DiskStorageAccountType)),
-					DiskType:               to.Ptr(parameters.DiskType),
-				},
-				AvailabilityZone: to.Ptr(parameters.AvailabilityZone),
-			},
-			AutoRepair: to.Ptr(parameters.AutoRepair),
-		},
-	}
-
-	if parameters.AutoScaling != nil {
-		nodePool.Properties.AutoScaling = &hcpsdk20260630preview.NodePoolAutoScaling{
-			Min: to.Ptr(parameters.AutoScaling.Min),
-			Max: to.Ptr(parameters.AutoScaling.Max),
-		}
-	} else {
-		nodePool.Properties.Replicas = to.Ptr(parameters.Replicas)
-	}
-
-	return nodePool
-}
-
-// CreateNodePoolAndWait20260630 creates a nodepool using the v20260630preview API and waits for completion.
-func CreateNodePoolAndWait20260630(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20260630preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-	nodePool hcpsdk20260630preview.NodePool,
-	timeout time.Duration,
-) (*hcpsdk20260630preview.NodePool, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout, fmt.Errorf("timeout '%f' minutes exceeded during CreateNodePoolAndWait20260630 for nodepool %s in cluster %s in resource group %s", timeout.Minutes(), nodePoolName, hcpClusterName, resourceGroupName))
-	defer cancel()
-	poller, err := nodePoolsClient.BeginCreateOrUpdate(ctx, resourceGroupName, hcpClusterName, nodePoolName, nodePool, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed starting nodepool creation %q for cluster %q in resourcegroup=%q: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	operationResult, err := poller.PollUntilDone(ctx, &runtime.PollUntilDoneOptions{
-		Frequency: StandardPollInterval,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed waiting for nodepool=%q for cluster %q in resourcegroup=%q to finish creating: %w", nodePoolName, hcpClusterName, resourceGroupName, err)
-	}
-
-	// Verify the LRO result body matches a fresh GET, per ARM LRO contract.
-	expect, err := GetNodePool20260630(ctx, nodePoolsClient, resourceGroupName, hcpClusterName, nodePoolName)
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return nil, fmt.Errorf("failed to get nodepool, caused by: %w, error: %w", context.Cause(ctx), err)
-		}
-		return nil, err
-	}
-	if err := checkOperationResult(expect, &operationResult.NodePool); err != nil {
-		return nil, err
-	}
-
-	return &operationResult.NodePool, nil
-}
-
-// GetNodePool20260630 retrieves a nodepool using the v20260630preview API.
-func GetNodePool20260630(
-	ctx context.Context,
-	nodePoolsClient *hcpsdk20260630preview.NodePoolsClient,
-	resourceGroupName string,
-	hcpClusterName string,
-	nodePoolName string,
-) (*hcpsdk20260630preview.NodePool, error) {
-	resp, err := nodePoolsClient.Get(ctx, resourceGroupName, hcpClusterName, nodePoolName, nil)
-	if err != nil {
-		return nil, err
-	}
-	return &resp.NodePool, nil
-}
-
-// BuildHCPClusterFromParams20251223 builds a v20251223preview HCP cluster from ClusterParams20251223
